cf_doorlight: kaldachom_stateHandlerA07 99.07->100 (unit -> 100.0)
#83(c): dropped the explicit (s32) cast on the float->alpha store - the
cast executed a spurious clrlwi after the fctiwz lwz; the direct float
assignment narrows at the stb with no mask, matching target.

+++ src/main/dll/cf_doorlight.c
@@ -122,7 +122,7 @@ int kaldachom_stateHandlerA07(int obj, int p)
     }
   }
   ((GameObject *)obj)->anim.alpha =
-      (s32)((lbl_803E3078 - ((GameObject *)obj)->anim.currentMoveProgress) * lbl_803E308C);
+      (lbl_803E3078 - ((GameObject *)obj)->anim.currentMoveProgress) * lbl_803E308C;
   return 0;
 }
